[wasm] add a test to tx_template
